# packages/api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
...
  "paths": {
...
    "/resumes/{id}/content": {
      "get": {
        "tags": [
          "resume-controller"
        ],
        "operationId": "content",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeContentResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "resume-controller"
        ],
        "operationId": "saveContent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeContent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeContentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs/{id}/applied": {
      "put": {
        "tags": [
          "job-controller"
        ],
        "operationId": "applied",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "job-controller"
        ],
        "operationId": "notApplied",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/billing/checkout": {
      "post": {
        "tags": [
          "billing-controller"
        ],
        "operationId": "checkout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutLinkView"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/billing/users/{userId}/adjustments": {
      "post": {
        "tags": [
          "admin-billing-controller"
        ],
        "operationId": "adjust",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdjustmentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/billing/ledger": {
      "get": {
        "tags": [
          "billing-controller"
        ],
        "operationId": "ledger",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerPage"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/billing/costs": {
      "get": {
        "tags": [
          "admin-billing-controller"
        ],
        "operationId": "costs",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CostReportResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Certification": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "issuer": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "date": {
            "type": "string",
            "pattern": "^\\d{4}(-(0[1-9]|1[0-2]))?$"
          }
        },
        "required": [
          "name"
        ]
      },
      "Contact": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "email": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0,
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$"
          },
          "phone": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "location": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            },
            "maxItems": 10,
            "minItems": 0
          }
        }
      },
      "Education": {
        "type": "object",
        "properties": {
          "institution": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "degree": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "field_of_study": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "start_date": {
            "type": "string",
            "pattern": "^\\d{4}(-(0[1-9]|1[0-2]))?$"
          },
          "end_date": {
            "type": "string",
            "pattern": "^\\d{4}(-(0[1-9]|1[0-2]))?$"
          }
        },
        "required": [
          "institution"
        ]
      },
      "Experience": {
        "type": "object",
        "properties": {
          "company": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "location": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "start_date": {
            "type": "string",
            "pattern": "^\\d{4}(-(0[1-9]|1[0-2]))?$"
          },
          "end_date": {
            "type": "string",
            "pattern": "^\\d{4}(-(0[1-9]|1[0-2]))?$"
          },
          "is_current": {
            "type": "boolean"
          },
          "bullets": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 2000,
              "minLength": 0
            },
            "maxItems": 25,
            "minItems": 0
          }
        },
        "required": [
          "company",
          "title"
        ]
      },
      "Link": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "url": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0,
            "pattern": "^(?i)https?://\\S+$"
          }
        },
        "required": [
          "url"
        ]
      },
      "Project": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 0
          },
          "url": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0,
            "pattern": "^(?i)https?://\\S+$"
          },
          "technologies": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 200,
              "minLength": 0
            },
            "maxItems": 25,
            "minItems": 0
          }
        },
        "required": [
          "name"
        ]
      },
      "ResumeContent": {
        "type": "object",
        "properties": {
          "schema_version": {
            "type": "integer",
            "format": "int32"
          },
          "contact": {
            "$ref": "#/components/schemas/Contact"
          },
          "headline": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "summary": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 0
          },
          "experience": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Experience"
            },
            "maxItems": 30,
            "minItems": 0
          },
          "education": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Education"
            },
            "maxItems": 15,
            "minItems": 0
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 100,
              "minLength": 0
            },
            "maxItems": 100,
            "minItems": 0
          },
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            },
            "maxItems": 20,
            "minItems": 0
          },
          "certifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Certification"
            },
            "maxItems": 20,
            "minItems": 0
          }
        }
      },
      "ParseWarning": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
      "ResumeContentResponse": {
        "type": "object",
        "properties": {
          "resumeId": {
            "type": "string",
            "format": "uuid"
          },
          "parseStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "PARSED",
              "FAILED"
            ]
          },
          "parseError": {
            "type": "string"
          },
          "versionNumber": {
            "type": "integer",
            "format": "int32"
          },
          "source": {
            "type": "string",
            "enum": [
              "UPLOAD",
              "EDIT",
              "TAILORED"
            ]
          },
          "content": {
            "$ref": "#/components/schemas/ResumeContent"
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParseWarning"
            }
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CheckoutRequest": {
        "type": "object",
        "properties": {
          "plan": {
            "type": "string",
            "maxLength": 30,
            "minLength": 0
          },
          "pack": {
            "type": "string",
            "maxLength": 60,
            "minLength": 0
          },
          "provider": {
            "type": "string",
            "maxLength": 20,
            "minLength": 0
          }
        },
        "required": [
          "provider"
        ]
      },
      "CheckoutLinkView": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
...
      "AdjustmentRequest": {
        "type": "object",
        "properties": {
          "idempotencyKey": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "delta": {
            "type": "number",
            "exclusiveMaximum": 0
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "delta",
          "idempotencyKey",
          "reason"
        ]
      },
      "AdjustmentResponse": {
        "type": "object",
        "properties": {
          "applied": {
            "type": "boolean"
          },
          "balance": {
            "type": "number"
          }
        }
      },
...
      "LedgerLine": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "delta": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          },
          "balanceAfter": {
            "type": "number"
          },
          "feature": {
            "type": "string"
          }
        }
      },
      "LedgerPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LedgerLine"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        }
      },
...
      "CostReportResponse": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "totals": {
            "$ref": "#/components/schemas/CostRowResponse"
          },
          "byFeature": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CostRowResponse"
            }
          },
          "byDay": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CostRowResponse"
            }
          },
          "byModel": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CostRowResponse"
            }
          },
          "byDayFeature": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CostRowResponse"
            }
          }
        }
      },
      "CostRowResponse": {
        "type": "object",
        "properties": {
          "day": {
            "type": "string",
            "format": "date"
          },
          "feature": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "calls": {
            "type": "integer",
            "format": "int64"
          },
          "failedCalls": {
            "type": "integer",
            "format": "int64"
          },
          "inputTokens": {
            "type": "integer",
            "format": "int64"
          },
          "outputTokens": {
            "type": "integer",
            "format": "int64"
          },
          "costUsd": {
            "type": "number"
          }
        }
      },
...
    }
  }
}